feat: void function calls
